+++ config.js
@@ -4,9 +4,6 @@ export const githubRepository =
 export const githubDefaultBranch = 'live';
 export const githubContribute = `${githubRepository}blob/${githubDefaultBranch}/CONTRIBUTING.md`;
 
-export const metaDescription =
-  'A quick and easy way to figure out what the best tool for your next project is, if it’s worth your time to migrate from one tool to another and how to adopt a best practice into your existing code base. Brought to you by web.dev';
-
 const now = new Date();
 export const buildDate = [
   now.getUTCFullYear(),
