+++ static-build/components/HeadMeta/index.tsx
@@ -11,7 +11,6 @@
  * limitations under the License.
  */
 import { h, Fragment } from 'preact';
-import SocialMeta from './social';
 import sharedStyles from 'css-bundle:static-build/shared/styles/index.css';
 
 import icoURL from 'asset-url:../../img/favicon.ico';
@@ -20,20 +19,40 @@ import png32URL from 'asset-url:../../img/favicon-32x32.png';
 import png16URL from 'asset-url:../../img/favicon-16x16.png';
 import safariURL from 'asset-url:../../img/safari-pinned-tab.svg';
 
-export default function HeadMeta() {
+import toolingReportSocial from 'asset-url:static-build/img/tooling-report-social.png';
+
+interface Props {
+  titleParts: string[];
+  description: string;
+}
+
+export default function HeadMeta({ titleParts, description }: Props) {
+  const title = [...titleParts, 'Tooling.Report'].join(' | ');
+
   return (
     <Fragment>
       <meta
         name="viewport"
         content="width=device-width,initial-scale=1,minimum-scale=1"
       />
+      <title>{title}</title>
+      <meta property="og:title" content={title} />
+      <meta name="description" content={description} />
+      <meta property="og:description" content={description} />
       <link rel="shortcut icon" href={icoURL} />
       <link rel="apple-touch-icon" sizes="180x180" href={appleIconURL} />
       <link rel="icon" type="image/png" sizes="32x32" href={png32URL} />
       <link rel="icon" type="image/png" sizes="16x16" href={png16URL} />
       <link rel="mask-icon" href={safariURL} />
       <meta name="color-scheme" content="dark light" />
-      <SocialMeta />
+      <meta property="og:type" content="article" />
+      <meta property="og:image" content={toolingReportSocial} />
+      <meta property="twitter:image" content={toolingReportSocial} />
+      <meta property="og:site_name" content="tooling.report" />
+      <meta name="twitter:site" content="@ChromiumDev" />
+      <meta name="twitter:creator" content="@ChromiumDev" />
+      <meta name="twitter:card" content="summary_large_image" />
+      <meta name="twitter:image:alt" content="tooling.report" />
       <link rel="stylesheet" href={sharedStyles} />
     </Fragment>
   );
